+++ merlin/main.py
@@ -190,7 +190,7 @@ def launch_workers(args):
     if args.worker_echo_only:
         print(status)
     else:
-        LOG.info(status)
+        LOG.debug(f"celery command: {status}")
 
 
 def purge_tasks(args):
@@ -552,7 +552,8 @@ def setup_argparse():
 
     # merlin info
     info = subparsers.add_parser(
-        "info", help="show pip and python versions and locations"
+        "info",
+        help="display info about the merlin configuration and the python configuration. Useful for debugging.",
     )
     info.set_defaults(func=print_info)