+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Updated the Merlin Sphinx web docs.
 - Changed the example workflows to use python3 instead of python.
+- Updated `merlin info` to lookup python3 and and pip3.
 - Altered user in Dockerfile and removed build tools.
 - MANIFEST.in now uses recursive-include.
 - Added author and author_email to setup.py.
@@ -28,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - Makefile targets `pull` and `update`.
 - Unneeded variables from `simple_chain.yaml`.
+- All `INFO`-level logger references to Celery.
 
 ### Changed
 - Updated the Merlin Sphinx web docs.